backend/requirements.txt
flask
flask-cors
python-dotenv
supabase
google-generativeai
pytest
python-docx
reportlab
pydantic
markdown2
